put hoverPoint and getTraceColor in separate files
